Add Loggability::SpecHelpers for setting up logging in tests.
